Moved WhatsApp function to whatsapp file
